requirements.txt
torch
torchvision
numpy
pandas
matplotlib
jupyter
scikit-learn
gradio
tqdm
seaborn
deeplake
flask
pillow
flask-cors
kornia